+++ WorldCities/WorldCities/src/app/cities/city-edit.component.ts
@@ -1,7 +1,9 @@
 import { Component , OnInit } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { ActivatedRoute, Router } from '@angular/router';
-import { FormGroup, FormControl, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
+import { FormGroup, FormControl, FormsModule, ReactiveFormsModule, Validators, AbstractControl, AsyncValidatorFn } from '@angular/forms';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { City } from './city';
 import { Country } from '../countries/country';
 import { environment } from './../../envrionments/environment';
@@ -46,7 +48,7 @@ export class CityEditComponent implements OnInit{
       lat: new FormControl('', Validators.required),
       lon: new FormControl('', Validators.required),
       countryId: new FormControl('', Validators.required)
-    });
+    }, null, this.isDupeCity());
     this.loadData();
   }
 
@@ -122,6 +124,19 @@ export class CityEditComponent implements OnInit{
     }
   } // end of OnSubmit
 
+  isDupeCity(): AsyncValidatorFn {
+    return (control: AbstractControl): Observable<{ [key: string]: any } | null> => {
+      var city = <City>{};
+      city.id = (this.id) ? this.id : 0;
+      city.name = this.form.controls['name'].value;
+      city.lat = +this.form.controls['lat'].value;
+      city.lon = +this.form.controls['lon'].value;
+      city.countryId = +this.form.controls['countryId'].value;
 
-
+      var url = environment.baseURL + 'api/Cities/IsDupecity';
+      return this.http.post<boolean>(url, city).pipe(map(result => {
+        return (result ? { isDupeCity: true } : null);
+      }));
+    }
+  }
 }