+++ WorldCities/WorldCitiesAPI/Controllers/CitiesController.cs
@@ -117,5 +117,17 @@ public class CitiesController : ControllerBase
     return Ok();
   }
 
+  [HttpPost]
+  [Route("IsDupeCity")]
+  public bool IsDupeCity(City city)
+  {
+    return _context.Cities.Any(e => 
+    e.Name == city.Name
+    && e.Lat == city.Lat
+    && e.Lon == city.Lon
+    && e.CountryId == city.CountryId
+    && e.Id != city.Id);
+  }
+
 }